fix: handle unhandled rejection in SES in code bundling
This commit addresses an unhandled rejection error encountered during the code bundling process. The issue originated from the "@endo/bundle-source" dependency, version "^2.1.1". The error, identified as SES_UNHANDLED_REJECTION, was caused by a TypeError related to assigning a value to the read-only constructor property of an object. The detailed error message is as follows:

TypeError: Cannot assign to read only property constructor of object [object Object].